Create login-form file
